Massive refactoring; implement mouse-eye tracking
While everything fits together fairly well,
I'm somewhat unhappy with the overall
structure of the Animation component --
it's way too humongous.

I'm uncertain about continuing to use React
but I'm not interested in writing another
reactive UI library.
---

The eye-tracking is fairly cool. Glitchy if
the mouse is too close. A fix would be to
simply stop tracking when the mouse is on
the character's face. I'll consider rewo-
rking that later.